Code clean up - reduce redundant code
